centralize period indexing tests
